Delete docs/weekly logs/W5 Team Log.md
Need to fix error when we merge this branch to master. This file will cause an issue, but it has already been fixed in master, development is actually the one behind. Master still has this file, so I'm not really deleting it, it is just rename in the master branch.