added create user php
created a create user php file that the readInput unity script will connect to create a new user. Also, it uses a salt and hash system for the user password instead of the way the login.php did it last week. Need to get Unity to show changes now, but the php is mostly done.