add DBManager.cs
Created new script for tracking if the user is logged in, as well as all of that user's stats. I don't know if the datecreated or wlratio variables will work right now however, since I wrote this and the php on my laptop (without access to my usual server and db that I used for the system before on my desktop at home.), so i couldn't test or see the database info (I forgot to upload it to github). Will check it after.