adding main menu script here
Adding main menu script to this folder so we can see it on the Branch (it is normally in the assests folder of the project, and it wasn't anywhere on github before this (maybe in my other repo). I also have made some serious changes to it, in order to display the user's information once they successfully login now. The account info page as well as the stats page actually display the correct info now which is cool. Only thing is that when the user successfully logins in, they don't get taken from the login or create menu and sent back to the main menu yet like they should.